backend/requirements.txt
APScheduler==3.9.1
Django==5.0
djangorestframework==3.14.0
dramatiq==1.15.0
django-dramatiq==0.11.6
drf-yasg==1.21.7
coreapi==2.3.3
channels-redis==4.3.0
channels==4.3.1
psycopg2-binary
marshmallow==3.14.1
marshmallow-dataclass==8.6.0
marshmallow-enum==1.5.1
PyYAML
prompt-toolkit==3.0.21
django-extensions==3.2.3
ptpython==3.0.20
pytest-django==4.7.0
mixer==7.2.0
Faker==12.0.0
daphne==4.2.1
shyaml==0.6.2
redis==6.4.0

pika==1.3.2

# JWT Authentication
djangorestframework-simplejwt==5.3.0
dj-rest-auth==5.0.2
django-allauth
Pillow==10.3.0
djangochannelsrestframework==1.1.0

# RSS parsing
feedparser==6.0.10
requests==2.31.0

# Security and validation packages
django-ratelimit==4.1.0
django-cors-headers==4.3.1
django-filter==23.5
django-axes==6.1.0

# Type checking
mypy==1.7.1
django-stubs==4.2.7
types-redis==4.6.0

# UI improvements
django-admin-interface==0.28.6
django-colorfield==0.11.0

# Testing and code quality
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
coverage==7.3.2

# AI and sentiment analysis
openai==1.106.1

